backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]

# ── Config / env management ────────────────────────────────────────────────────
pydantic-settings
python-dotenv

# ── HTTP client (AbuseIPDB + Cloudflare API calls) ─────────────────────────────
httpx

# ── Database ───────────────────────────────────────────────────────────────────
SQLAlchemy[asyncio]
asyncpg
alembic

# ── Redis ──────────────────────────────────────────────────────────────────────
redis[asyncio]

# ── GeoIP (added in Step 4) ────────────────────────────────────────────────────
geoip2

# ── ML (added in Step 5) ───────────────────────────────────────────────────────
scikit-learn
joblib
pandas
numpy

# ── Scheduler / cron jobs ──────────────────────────────────────────────────────
apscheduler
